1. modify ftl formatter 2. Change valid to Byte 3. add some common config

+++ src/main/java/com/srct/service/config/db/DataSourceCommonConstant.java
@@ -15,9 +15,9 @@ package com.srct.service.config.db;
 public class DataSourceCommonConstant {
 
     // Common
-    public static final Integer DATABASE_COMMON_INVALID = 0;
+    public static final Byte DATABASE_COMMON_INVALID = 0;
 
-    public static final Integer DATABASE_COMMON_VALID = 1;
+    public static final Byte DATABASE_COMMON_VALID = 1;
 
-    public static final Integer DATABASE_COMMON_IGORE_VALID = -1;
+    public static final Byte DATABASE_COMMON_IGORE_VALID = -1;
 }
